swap: move swap API code into api.go file
